pyproject.toml
[tool.poetry]
name = "sidrograd"
version = "0.1.0"
description = ""
authors = ["pekshev"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.23"
djangorestframework = "3.14.0"
django-filter = "23.3"
django-cors-headers = "4.3.1"
pillow = "10.1.0"
pytils = "0.4.1"
tzdata = "2023.3"
six = "1.16.0"
python-dateutil = "2.8.2"
markdown = "3.5.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
